pyproject.toml
[tool.poetry]
name = "repl_python3_Locus-Verify"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
Flask = "^2.0.2"
Flask-Discord = "^0.1.64"
python-dotenv = "^0.19.2"
disnake = "^2.3.0"
Flask-xCaptcha = "^0.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
